[FIX] mail: prevent traceback error when deleting quotation attachment
Issue:
A traceback error is raised when deleting the quotation attachment before sending it via email.

Steps to reproduce:
	- go to Sales.
	- open a quotation.
	- click "Send by Email."
	- delete the attachment.

A traceback error occurs.

opw-4486182

closes odoo/odoo#194593

+++ addons/mail/static/src/core/web/mail_composer_attachment_list.js
@@ -20,7 +20,9 @@ export class MailComposerAttachmentList extends Many2ManyBinaryField {
     async onFileRemove(fileId) {
         super.onFileRemove(fileId);
         const attachment = this.mailStore.Attachment.get(fileId);
-        await this.attachmentUploadService.unlink(attachment);
+        if (attachment) {
+            await this.attachmentUploadService.unlink(attachment);
+        }
     }
 }
